Refactor apps skeleton
Refactor apps skeleton
