+++ backend/permissions/patient_permission.py
@@ -1,0 +1,19 @@
+from rest_framework import permissions
+
+
+class PatientPermission(permissions.BasePermission):
+    def has_permission(self, request, view):
+        if request.method == 'OPTIONS':
+            return True
+        if view.action in ['list', 'retrieve', 'diagnoses', 'medical_cards']:
+            return request.user.is_authenticated
+        else:
+            return False
+
+    def has_object_permission(self, request, view, obj):
+        if request.method == 'OPTIONS':
+            return True
+        if view.action in ['list', 'retrieve', 'diagnoses', 'medical_cards']:
+            return request.user.is_authenticated
+        else:
+            return False
