+++ backend/permissions/registry_permission.py
@@ -1,0 +1,37 @@
+from rest_framework import permissions
+
+
+class DiagnosisRegistryPermission(permissions.BasePermission):
+    def has_permission(self, request, view):
+        if request.method == 'OPTIONS':
+            return True
+        if view.action in ['list', 'create', 'retrieve', 'update', 'destroy']:
+            return request.user.is_authenticated
+        else:
+            return False
+
+    def has_object_permission(self, request, view, obj):
+        if request.method == 'OPTIONS':
+            return True
+        if view.action in ['list', 'create', 'retrieve', 'update', 'destroy']:
+            return request.user.is_authenticated
+        else:
+            return False
+
+
+class DiagnosisRegistryItemPermission(permissions.BasePermission):
+    def has_permission(self, request, view):
+        if request.method == 'OPTIONS':
+            return True
+        if view.action in ['list', 'create', 'retrieve', 'update', 'destroy']:
+            return request.user.is_authenticated
+        else:
+            return False
+
+    def has_object_permission(self, request, view, obj):
+        if request.method == 'OPTIONS':
+            return True
+        if view.action in ['list', 'create', 'retrieve', 'update', 'destroy']:
+            return request.user.is_authenticated
+        else:
+            return False
