--- backend/routers.py
@@ -1,73 +1,0 @@
-from rest_framework import routers
-
-
-# class EquipmentRouter(routers.SimpleRouter):
-#     routes = [
-#         routers.Route(url=r'^{prefix}$',
-#                       mapping={'get': 'list', 'post': 'create'},
-#                       name='{basename}-list',
-#                       detail=False,
-#                       initkwargs={'suffix': 'List'}),
-#         routers.Route(url=r'^{prefix}/{lookup}$',
-#                       mapping={'get': 'retrieve', 'put': 'update', 'delete': 'destroy'},
-#                       name='{basename}-detail',
-#                       detail=True,
-#                       initkwargs={'suffix': 'Detail'})
-#     ]
-#
-#
-class UserRouter(routers.SimpleRouter):
-    routes = [
-        routers.Route(
-            url=r'^{prefix}/register$',
-            mapping={'post': 'create'},
-            name='{basename}-list',
-            detail=False,
-            initkwargs={'suffix': 'List'}
-        ),
-        routers.Route(
-            url=r'^{prefix}/profile$',
-            mapping={'get': 'retrieve'},
-            name='{basename}-detail',
-            detail=True,
-            initkwargs={'suffix': 'Detail'}
-        ),
-    ]
-
-
-class OnlyListRouter(routers.SimpleRouter):
-    routes = [
-        routers.Route(
-            url=r'^{prefix}$',
-            mapping={'get': 'list'},
-            name='{basename}-list',
-            detail=False,
-            initkwargs={'suffix': 'List'}
-        ),
-    ]
-
-
-class ListRetrieveRouter(routers.SimpleRouter):
-    routes = [
-        routers.Route(
-            url=r'^{prefix}$',
-            mapping={'get': 'list'},
-            name='{basename}-list',
-            detail=False,
-            initkwargs={'suffix': 'List'}
-        ),
-        routers.Route(
-            url=r'^{prefix}/{lookup}$',
-            mapping={'get': 'retrieve'},
-            name='{basename}-detail',
-            detail=True,
-            initkwargs={'suffix': 'Detail'}
-        ),
-        routers.DynamicRoute(
-            url=r'^{prefix}/{lookup}/{url_path}$',
-            name='{basename}-{url_name}',
-            detail=True,
-            initkwargs={}
-        ),
-    ]
-
